Update README for .NET 9 and include in solution
The README.md file has been updated to indicate that the project is now ported to .NET 9, replacing the previous reference to .NET 8. Additionally, the README.md file has been added to the solution items in the Skoruba.AuditLogging.sln file to ensure it is part of the project structure.